divide up toolwidgets
