+++ meerk40t/gui/scene/scenepanel.py
@@ -1,4 +1,3 @@
-
 import wx
