+++ meerk40t/gui/scene/scenespacewidget.py
@@ -1,4 +1,8 @@
-from meerk40t.gui.scene.sceneconst import HITCHAIN_DELEGATE_AND_HIT, RESPONSE_CHAIN, RESPONSE_CONSUME
+from meerk40t.gui.scene.sceneconst import (
+    HITCHAIN_DELEGATE_AND_HIT,
+    RESPONSE_CHAIN,
+    RESPONSE_CONSUME,
+)
 from meerk40t.gui.scene.widget import Widget
 from meerk40t.svgelements import Matrix, Viewbox
