+++ meerk40t/gui/scene/widget.py
@@ -1,11 +1,21 @@
-
 from math import isinf, isnan
 
 import wx
 
-from meerk40t.gui.scene.sceneconst import ORIENTATION_RELATIVE, HITCHAIN_DELEGATE, RESPONSE_CHAIN, \
-    ORIENTATION_ABSOLUTE, ORIENTATION_NO_BUFFER, BUFFER, ORIENTATION_MODE_MASK, ORIENTATION_GRID, ORIENTATION_DIM_MASK, \
-    ORIENTATION_VERTICAL, ORIENTATION_HORIZONTAL, ORIENTATION_CENTERED
+from meerk40t.gui.scene.sceneconst import (
+    BUFFER,
+    HITCHAIN_DELEGATE,
+    ORIENTATION_ABSOLUTE,
+    ORIENTATION_CENTERED,
+    ORIENTATION_DIM_MASK,
+    ORIENTATION_GRID,
+    ORIENTATION_HORIZONTAL,
+    ORIENTATION_MODE_MASK,
+    ORIENTATION_NO_BUFFER,
+    ORIENTATION_RELATIVE,
+    ORIENTATION_VERTICAL,
+    RESPONSE_CHAIN,
+)
 from meerk40t.gui.zmatrix import ZMatrix
 from meerk40t.svgelements import Matrix
