+++ meerk40t/gui/wxmscene.py
@@ -12,18 +12,17 @@ from meerk40t.gui.scenewidgets.laserpathwidget import LaserPathWidget
 from meerk40t.gui.scenewidgets.rectselectwidget import RectSelectWidget
 from meerk40t.gui.scenewidgets.reticlewidget import ReticleWidget
 from meerk40t.gui.scenewidgets.selectionwidget import SelectionWidget
-from meerk40t.gui.scenewidgets.toolwidgets import (
-    CircleTool,
-    DrawTool,
-    EllipseTool,
-    PolygonTool,
-    PolylineTool,
-    RectTool,
-    RelocateTool,
-    TextTool,
-    ToolContainer,
-    VectorTool,
-)
+from meerk40t.gui.toolwidgets.toolcircle import CircleTool
+from meerk40t.gui.toolwidgets.toolcontainer import ToolContainer
+from meerk40t.gui.toolwidgets.tooldraw import DrawTool
+from meerk40t.gui.toolwidgets.toolellipse import EllipseTool
+from meerk40t.gui.toolwidgets.toolpolygon import PolygonTool
+from meerk40t.gui.toolwidgets.toolpolyline import PolylineTool
+from meerk40t.gui.toolwidgets.toolrect import RectTool
+from meerk40t.gui.toolwidgets.toolrelocate import RelocateTool
+from meerk40t.gui.toolwidgets.tooltext import TextTool
+from meerk40t.gui.toolwidgets.toolvector import VectorTool
+
 from meerk40t.gui.wxutils import get_key_name
 from meerk40t.kernel import signal_listener
 from meerk40t.svgelements import Angle, Length
